Move players back to the board
